Polyfill the incremental generator ForAttributeWithMetadataName from roslyn (for LoggingGenerator).
