+++ src/libraries/Microsoft.Extensions.Logging.Abstractions/gen/LoggerMessageGenerator.Parser.cs
@@ -18,7 +18,7 @@ namespace Microsoft.Extensions.Logging.Generators
     {
         internal sealed class Parser
         {
-            private const string LoggerMessageAttribute = "Microsoft.Extensions.Logging.LoggerMessageAttribute";
+            internal const string LoggerMessageAttribute = "Microsoft.Extensions.Logging.LoggerMessageAttribute";
 
             private readonly CancellationToken _cancellationToken;
             private readonly Compilation _compilation;
@@ -33,33 +33,6 @@ namespace Microsoft.Extensions.Logging.Generators
 
             internal static bool IsSyntaxTargetForGeneration(SyntaxNode node) =>
                 node is MethodDeclarationSyntax m && m.AttributeLists.Count > 0;
-
-            internal static ClassDeclarationSyntax? GetSemanticTargetForGeneration(GeneratorSyntaxContext context)
-            {
-                var methodDeclarationSyntax = (MethodDeclarationSyntax)context.Node;
-
-                foreach (AttributeListSyntax attributeListSyntax in methodDeclarationSyntax.AttributeLists)
-                {
-                    foreach (AttributeSyntax attributeSyntax in attributeListSyntax.Attributes)
-                    {
-                        IMethodSymbol attributeSymbol = context.SemanticModel.GetSymbolInfo(attributeSyntax).Symbol as IMethodSymbol;
-                        if (attributeSymbol == null)
-                        {
-                            continue;
-                        }
-
-                        INamedTypeSymbol attributeContainingTypeSymbol = attributeSymbol.ContainingType;
-                        string fullName = attributeContainingTypeSymbol.ToDisplayString();
-
-                        if (fullName == LoggerMessageAttribute)
-                        {
-                            return methodDeclarationSyntax.Parent as ClassDeclarationSyntax;
-                        }
-                    }
-                }
-
-                return null;
-            }
 
             /// <summary>
             /// Gets the set of logging classes containing methods to output.
